refactor: decouple EmbPattern from readers/writers (#123)
In the reader and writer implementations we want to add EmbPattern as type
hints, however when imported relatively this lead to a circular import since
EmbPattern.py imported all the readers and writers.

This commit resolves this circular dependency by moving the read and write
methods that select the writer depending on the file extension to __init__.py
(as well as supported_formats and the format-specific read and write methods).